Remove binding of haskell-mode-jump-to-def-or-tag
This binding is configured by default these days. And the mode hook code I removed would have had the effect of binding the key every time a haskell file was opened: it would have been sufficient to just put the `(define-key ...)` form at the top level, or in an `(eval-after-load 'haskell-mode ...)` wrapper.